Use single precision

+++ vgn/src/vgn/utils/__init__.py
@@ -17,7 +17,7 @@ def load_image(fname):
 
 def voxel_grid_to_array(voxel_grid, resolution):
     """Convert an Open3D voxel grid to a contiguous 3D numpy array."""
-    v = np.zeros(shape=(resolution, resolution, resolution))
+    v = np.zeros(shape=(resolution, resolution, resolution), dtype=np.float32)
     for voxel in voxel_grid.voxels:
         i, j, k = voxel.grid_index
         v[i, j, k] = voxel.color[0]
